requirements.txt
pandas
azure-storage-blob
sqlalchemy
pyodbc
python-dotenv
